% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_pwycommons_sif.R
\name{clean_pwycommons_sif}
\alias{clean_pwycommons_sif}
\title{Clean Pathway Commons SIF file}
\usage{
clean_pwycommons_sif(pc.sif, rm.ids = "CHEBI:15377")
}
\arguments{
\item{pc.sif}{SIF file from Pathway Commons}

\item{rm.ids}{vector of analyte IDs to remove. Default \code{CHEBI:15377} corresponds to water.}
}
\value{
SIF file
}
\description{
Apply a few curation steps to "clean" Pathway Commons Simple Interaction Format (SIF) file, 
based on manual curation of Pathway Commons 8.
}
