% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_bar.R
\name{color_bar}
\alias{color_bar}
\title{Plot colorbar}
\usage{
color_bar(col, lev, signif.dig = 2)
}
\arguments{
\item{col}{Color palette vector. First value can be NA. \code{NA} is transformed to white, so \code{col} cannot have
\code{"white"} or its hex version \code{"#ffffff"}.}

\item{lev}{Vector of numeric levels corresponding to colors. The min and max of this vector is used to define numeric 
range. Cannot have NAs.}

\item{signif.dig}{Number of significant digits to use for colorbar axis labels.}
}
\value{
Invibly, \code{col}.
}
\description{
Plot colorbar. There are many functions for this in other pacakges, but they do not appear to handle NAs.
}
\details{
Modified from \url{https://aurelienmadouasse.wordpress.com/2012/01/13/legend-for-a-continuous-color-scale-in-r/}.
}
