% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_kernel.R
\name{diag_kernel}
\alias{diag_kernel}
\title{Make diagonal kernel matrix}
\usage{
diag_kernel(object.rownames, Gmat.rownames)
}
\arguments{
\item{object.rownames}{Rownames of \code{object}. Must be non-duplicated and non-empty.}

\item{Gmat.rownames}{Rownames of \code{Gmat}. Must be non-duplicated and non-empty.}
}
\value{
A diagonal matrix.
}
\description{
Make diagonal kernel matrix to run Pants without smoothing.
}
\details{
\code{object.rownames} and \code{Gmat.rownames} should have some overlap.
}
