% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_graph.R
\name{expand_graph}
\alias{expand_graph}
\title{Expand graph to match features in pathway matrix}
\usage{
expand_graph(gr, Gmat)
}
\arguments{
\item{gr}{Interaction network of class \code{igraph}.}

\item{Gmat}{Pathway membership matrix, can be sparse matrix from package \code{Matrix}.}
}
\value{
Interaction network of class \code{igraph}.
}
\description{
Expand graph to match features in pathway matrix, so that maintain these features in 
\code{\link{pants}}.
}
\details{
Nodes in \code{Gmat} but not \code{gr} are added as isolated nodes to \code{gr} so that their scores
can be incorporated in \code{\link{pants}} and plotted in \code{\link{plot_pwy}}.
}
