% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_kernel.R
\name{expand_kernel}
\alias{expand_kernel}
\title{Expand kernel to match features in object and pathway matrix}
\usage{
expand_kernel(ker, Gmat, object.feat.nms)
}
\arguments{
\item{ker}{sparse matrix (of class Matrix)}

\item{Gmat}{sparse matrix (of class Matrix)}

\item{object.feat.nms}{feature names from object.}
}
\value{
sparse kernel matrix (of class Matrix)
}
\description{
Expand kernel to match features in both object and pathway matrix, so that maintain these features in 
\code{\link{pants}}.
}
