% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezpermutations.R
\name{ezpermutations}
\alias{ezpermutations}
\title{Resample sample labels for permutation testing}
\usage{
ezpermutations(xx, nperm, freq = length(xx) - length(unique(xx)) + 1)
}
\arguments{
\item{xx}{Vector of labels or indices to resample. Must be of length at least 2.}

\item{nperm}{Number of resamples to return.}

\item{freq}{Integer number of times to repeat each element of \code{names(table(xx))} within each resample.
Can be a vector where \code{length(freq)==length(table(xx))} or of length 1, which is recycled into a vector.
Represents resampling without replacement if it's \code{table(xx)}. For resampling with replacement (bootstrapping),
\code{length(xx)-length(table(xx))+1} allows each element of \code{unique(xx)} to appear at least once 
in each resample.}
}
\value{
Data frame whose columns are same type as \code{xx}
}
\description{
Resample unique sample or group labels (unique elements of \code{xx}) into \code{nperm} data frame columns of length
\code{length(xx)} for permutation testing. The columns are not duplicates of each other, 
do not contain \code{xx}, and have each element of \code{unique(xx)} to appear at least once.
Calculations are done with \pkg{RcppAlgos}.
}
\details{
Number of returned resamples may be less than \code{nperm} if not enough unique resamples are available.
}
