% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmt2Gmat.R
\name{gmt2Gmat}
\alias{gmt2Gmat}
\title{Transform list from GMT into pathway matrix}
\usage{
gmt2Gmat(gmt.lst, sc = FALSE)
}
\arguments{
\item{gmt.lst}{List returned by \code{\link[ezlimma]{read_gmt}}.}

\item{sc}{Logical indicating to scale output Matrix by columns or not.}
}
\value{
A numeric Matrix of features (e.g. proteins) as rows and pathways as columns, indicating pathway membership
}
\description{
Transform list from reading in GMT using \code{ezlimma::read_gmt} into pathway matrix.
}
