% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_colorbar.R
\name{legend_colorbar}
\alias{legend_colorbar}
\title{Plot colorbar legend}
\usage{
legend_colorbar(col, lev)
}
\arguments{
\item{col}{Color palette vector}

\item{lev}{Levels corresponding to colors}
}
\description{
Plot colorbar legend
}
\details{
Modified from \url{https://aurelienmadouasse.wordpress.com/2012/01/13/legend-for-a-continuous-color-scale-in-r/}.
}
