% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map2color.R
\name{map2color}
\alias{map2color}
\title{Map values to colors}
\usage{
map2color(x, pal, limits = NULL)
}
\arguments{
\item{x}{Values to map.}

\item{pal}{Color palette.}

\item{limits}{A vector of size 2.}
}
\description{
Map values to colors in a palette.
}
\details{
From \href{https://stackoverflow.com/questions/15006211/how-do-i-generate-a-mapping-from-numbers-to-colors-in-r}{stack overflow}.
}
