% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_pow.R
\name{mat_pow}
\alias{mat_pow}
\title{Calculate power of a matrix}
\usage{
mat_pow(xx, pow)
}
\arguments{
\item{xx}{a square matrix.}

\item{pow}{positive integer power to raise the matrix to.}
}
\value{
a matrix.
}
\description{
Calculate matrix to an integer power.
}
\details{
This is a simple, recursive function to use integer powers when calculating p-step
random walk kernels.
}
