% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_mats.R
\name{match_mats}
\alias{match_mats}
\title{Expand kernel to match features in object and pathway matrix}
\usage{
match_mats(score.mat, ker, Gmat, score.impute = 0)
}
\arguments{
\item{score.mat}{A matrix-like object with scores from original data and possibly permutations with rows corresponding 
to features and columns to simulations.}

\item{ker}{sparse matrix (of class Matrix).}

\item{Gmat}{Feature-by-pathway inclusion matrix, indicating which features are in which pathways.}

\item{score.impute}{Value to impute missing scores. If zero, \code{score.mat} is sparser, saving computing time.}
}
\value{
A list with elements \code{score.mat}, \code{ker}, and \code{Gmat} after matching.
}
\description{
Expand kernel to match features in both object and pathway matrix, so that maintain these features in 
\code{\link{pants}}.
}
\details{
There must be some overlap between the rownames of \code{score.mat}, \code{ker}, & \code{Gmat}.
}
