% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchGmat2Ker.R
\name{matchGmat2Ker}
\alias{matchGmat2Ker}
\title{Match pathway matrix features to kernel matrix features}
\usage{
matchGmat2Ker(Gmat, ker)
}
\arguments{
\item{Gmat}{sparse matrix (of class Matrix)}

\item{ker}{sparse matrix (of class Matrix)}
}
\value{
sparse Gmat matrix (of class Matrix)
}
\description{
Match pathway matrix (Gmat) features to kernel matrix features, so that we can multiply these matrices
later.
}
