% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbor_nms.R
\name{neighbor_nms}
\alias{neighbor_nms}
\title{Get names of nodes' neighbors}
\usage{
neighbor_nms(graph, nodes)
}
\arguments{
\item{graph}{A graph object from package \code{igraph}}

\item{nodes}{Node names in the graph}
}
\description{
Get node names of nodes' direct neighbors
}
