% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_binom_n.R
\name{opt_binom_n}
\alias{opt_binom_n}
\title{Find minimal binomial N that ensures \code{ezpermutations} obtains \code{nperm} unique permutations}
\usage{
opt_binom_n(p, nperm)
}
\arguments{
\item{p}{Probablity a resample is successful.}

\item{nperm}{Number of resamples to return.}
}
\description{
Find minimal binomial N that ensures \code{ezpermutations} obtains \code{nperm} unique permutations. 
Let \eqn{X~binom(N, p=prob.nonrej)}, so \eqn{P(X>nperm)<\epsilon} <--> \eqn{Np-8\sqrt{Npq}-nperm=0}
<--> \eqn{\sqrt{N}=(\frac{8\sqrt{pq}+\sqrt{64pq + 4 \cdot p \cdot nperm}}{2p})} by quadratic formula in \eqn{\sqrt{N}}.
}
