% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_ecdf.R
\name{p_ecdf}
\alias{p_ecdf}
\title{Estimate p-value from simulations}
\usage{
p_ecdf(eval.v, score.mat, alternative = c("greater", "two.sided",
  "less"))
}
\arguments{
\item{eval.v}{A vector of values where the cumulative distribution function should be evaluated.}

\item{score.mat}{A matrix-like object with scores from original data and possibly permutations with rows corresponding 
to features and columns to simulations.}

\item{alternative}{A character string specifying the alternative hypothesis.}
}
\value{
A matrix with two columns containing z-scores (larger is more significant) & p-values with 
\code{nrow = length(eval.v)}.
}
\description{
Estimate p-value by comparing a score to its permutations.
}
\details{
It's checked that \code{rownames(score.mat)==names(eval.v)}. If a p-value is 1, it results in a z-score of 
-Inf, which can produce an error in downstream analysis, so we calculate a new p-value as 1 - 10^(-6).
}
\references{
Phipson B, Smyth GK. Permutation P-values should never be zero: calculating exact P-values when 
permutations are randomly drawn. Stat Appl Genet Mol Biol. 2010;9:Article39. doi: 10.2202/1544-6115.1585.
}
