% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_smooth_ecdf.R
\name{p_smooth_ecdf}
\alias{p_smooth_ecdf}
\title{Estimate p-value from a smooth density}
\usage{
p_smooth_ecdf(eval.point, scores, lower.tail = TRUE, smooth.fam = c("kde",
  "norm", "exp"))
}
\arguments{
\item{eval.point}{A value where the cumulative distribution function should be evaluated}

\item{scores}{A vector of values from the distribution}

\item{lower.tail}{Logical; if TRUE (default), probabilities are P[X ≤ x] otherwise, P[X > x].}

\item{smooth.fam}{One of "kde" for kernel density estimation, "norm" for the normal distribution or "exp" for
exponential distribution.}
}
\description{
Estimate p-value from a density smoothed with a non-parametric kernel density estimator or a known distribution
}
