% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pants_mediation.R
\name{pants_mediation}
\alias{pants_mediation}
\title{Pathway Analysis via Network Smoothing using mediation}
\usage{
pants_mediation(object, exposure.v, phenotype.v, ker, Gmat, nperm = 10^4,
  ret.null.mats = FALSE, verbose = TRUE)
}
\arguments{
\item{object}{A matrix-like data object containing log-ratios or log-expression values for a
series of samples, with rows corresponding to features and columns to samples.}

\item{exposure.v}{A numeric or character vector of exposures the same length as number of samples in \code{object}.
If the vector is named, the names must match the column names of \code{object}.}

\item{phenotype.v}{A vector of numeric phenotypes the same length as number of samples in \code{object}. If the 
vector is named, the names must match the column names of \code{object}.}

\item{ker}{The Laplacian kernel matrix.}

\item{Gmat}{The feature by pathway inclusion matrix, indicating which features are in which pathways.}

\item{nperm}{Number of permutations to perform to evaluate significance of pathways.}

\item{ret.null.mats}{If TRUE, return matrices with null distributions for features and pathways.}

\item{verbose}{Logical indicating if the permutation number should be output every 500 permutations.}
}
\value{
List of length at least two, with elements \code{pwy.stats} and \code{feature.stats}, each containing a data frame
of statistics.
}
\description{
PANTS algorithm for mediation pathway analysis (and optionally integration) via network smoothing
}
\details{
Scores for features in the kernel but not in the data are assigned a z-score of zero by default for sparsity.
Scores for features and pathways are compared to null scores, which are generated by permuting the columns of 
\code{object} and rerunning the algorithm.
}
