% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pc9.R
\docType{data}
\name{pc9}
\alias{pc9}
\title{Pathway Commons v9 in Simple Interaction Format (SIF)}
\format{A data frame with >1.5 million rows and 3 variables, representing: physical_entity_id <relationship type> physical_entity_id.}
\source{
\url{http://www.pathwaycommons.org/archives/PC2/v9/PathwayCommons9.All.hgnc.sif.gz}
}
\usage{
pc9
}
\description{
Pathway Commons v9 in SIF. SIF is described at \url{http://www.pathwaycommons.org/pc/sif_interaction_rules.do}.
Genes are represented by gene symbol (hgnc) and metabolites by CHEBI ID. All of the data provided by Pathway
Commons is free. In particular, Pathway Commons distributes pathway information with the intellectual property
restrictions of the source database; only databases that are freely available for academics are included.
}
\keyword{datasets}
