% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_ntop.R
\name{select_ntop}
\alias{select_ntop}
\title{Select top drivers for a pathway}
\usage{
select_ntop(score.v, Gmat, pwy, ker, alternative = c("two.sided", "less",
  "greater"), ntop = 3)
}
\arguments{
\item{score.v}{Named vector of scores of features to select driver nodes.}

\item{Gmat}{Binary feature (e.g. gene) by pathway inclusion matrix, indicating which features are in which pathways.}

\item{pwy}{Pathway, must be a column name of \code{Gmat}.}

\item{ker}{Laplacian kernel matrix.}

\item{alternative}{Alternative hypothesis; must be one of \code{"two.sided"}, \code{"greater"} or \code{"less"}. 
\code{"greater"} corresponds to positive association, \code{"less"} to negative association.}

\item{ntop}{Number of top features that most impact the pathway to include.}
}
\value{
Data frame with \code{ntop} rows ordered by impact & 3 columns: \code{node} with node names; 
\code{impact} with impact values; \code{in.pwy} with logicals if node is in \code{pwy}.
}
\description{
Select \code{ntop} top drivers for a pathway.
}
