% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pants_mediation.R
\name{sim_pants_mediation}
\alias{sim_pants_mediation}
\title{Simulate \code{pants} with mediation to estimate size & power}
\usage{
sim_pants_mediation(Gmat, exposure, effect.v = c(0, 0.2), alpha = 0.05,
  nsim = 10^3, nperm = 10^3, seed = 1, verbose = TRUE,
  ker = NULL, ncores = 1)
}
\arguments{
\item{Gmat}{Binary feature (e.g. gene) by pathway inclusion matrix, indicating which features are in which pathways.}

\item{exposure}{Numeric vector or matrix of exposures. Ignored if \code{type!="mediation"}.}

\item{effect.v}{Numeric vector of log fold-changes or percent of phentotypes to add.}

\item{alpha}{Alpha level.}

\item{nsim}{Number of simulations.}

\item{nperm}{Number of sample permutations to evaluate significance of pathways.}

\item{seed}{Random seed.}

\item{verbose}{Logical; should the number of simulations be printed every 100 simulations?}

\item{ker}{Laplacian kernel matrix representing the interaction network.}

\item{ncores}{Number of cores to use for parallel computing. You can detect how many are available for your system
using \code{\link[parallel]{detectCores}}.}
}
\description{
Simulate \code{pants} with mediation to estimate size (effect=0) & power (effect>0).
}
\seealso{
\code{\link{sim_pants}}
}
