% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMPDB2Gmat.R
\name{SMPDB2Gmat}
\alias{SMPDB2Gmat}
\title{Transform SMPDB protein and/or metabolite pathway data frames into a pathway matrix}
\usage{
SMPDB2Gmat(smpdb.prot, smpdb.met, exclude.pwy.subj = NA, sc = FALSE)
}
\arguments{
\item{smpdb.prot}{data frame of proteins per pathway from SMPDB}

\item{smpdb.met}{data frame of metabolites per pathway from SMPDB}

\item{exclude.pwy.subj}{vector of subjects (or types) of pathways to exclude}
}
\value{
A numeric Matrix of features (proteins and/or metabolites) as rows and pathways as columns, indicating pathway membership
}
\description{
Transform SMPDB protein and/or metabolite pathway data frames, read in as CSV files from http://smpdb.ca/downloads,
into a pathway matrix that \code{PANTS} can use to compute pathway scores.
}
