% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_smpdb_gmat.R
\docType{data}
\name{smpdb_gmat}
\alias{smpdb_gmat}
\title{Pathway inclusion matrix (Gmat) constructed from SMPDB.}
\format{A Matrix of features (proteins/metabolites) by Pathway names. Zero entries indicate the feature is not in the pathway,
whereas non-zero entries indicates otherwise.}
\usage{
smpdb_gmat
}
\description{
Pathway inclusion matrix (Gmat) constructed from http://smpdb.ca/downloads. These downloads are from 2014-2016. 
New downloads from early 2018 have very different pathways, with many nearly identical fatty acid biosynthesis 
pathways, so this object has not been updated with them. Code to reproduce this dataset from these downloads is at
https://github.com/jdreyf/PANTS/blob/master/data-raw/smpdb_gmat.R.
}
\keyword{datasets}
