% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_gmat.R
\name{subset_gmat}
\alias{subset_gmat}
\title{Subset pathway inclusion matrix based on size}
\usage{
subset_gmat(object, Gmat, min.nfeats)
}
\arguments{
\item{object}{Matrix-like data object containing log-ratios or log-expression values, with rows corresponding to 
features (e.g. genes) and columns to samples. Must have rownames that are non-duplicated and non-empty.}

\item{Gmat}{Binary feature (e.g. gene) by pathway inclusion matrix, indicating which features are in which pathways.}

\item{min.nfeats}{Minimum number of features (e.g. genes) needed in a gene set for testing.}
}
\value{
List of two: \code{Gmat} and vector \code{nfeats.per.pwy} with number of features per pathway.
}
\description{
Subset pathway inclusion matrix based on number of measured features in pathway.
}
