% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_kernel.R
\name{diag_kernel}
\alias{diag_kernel}
\title{Make diagonal kernel matrix}
\usage{
diag_kernel(object, Gmat)
}
\arguments{
\item{object}{Matrix-like data object containing log-ratios or log-expression values, with rows corresponding to 
features (e.g. genes) and columns to samples. Must have rownames that are non-duplicated and non-empty.}

\item{Gmat}{Binary feature (e.g. gene) by pathway inclusion matrix, indicating which features are in which pathways.}
}
\value{
A diagonal matrix.
}
\description{
Make diagonal kernel matrix to run Pants without smoothing.
}
\details{
\code{rownames(object)} and \code{rownames(Gmat)} should have some overlap.
}
