% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeList2graph.R
\name{edgeList2graph}
\alias{edgeList2graph}
\title{Transform a data frame edge list into an \code{igraph} graph}
\usage{
edgeList2graph(edge.lst)
}
\arguments{
\item{edge.lst}{Edge-list}
}
\value{
An \code{igraph} graph
}
\description{
Transforms an edge list into an \code{igraph} graph. The edge list can come from
columns 1 and 3 of a SIF (Simple Interaction Format) file, or can otherwise
contain pairs of nodes connected in the graph.
}
