% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezpermutations.R
\name{ezpermutations}
\alias{ezpermutations}
\title{Permute labels or indices for permutation testing}
\usage{
ezpermutations(xx, nperm, freq = length(xx) - 2)
}
\arguments{
\item{xx}{Vector of unique labels or indices to permute. Must be of length at least 3.}

\item{nperm}{Number of resamples to return.}

\item{freq}{Integer number of times to repeat any element of \code{xx} within each permutation. \code{freq=1}
implies sampling without replacement; \code{freq>1} implies sampling with replacement (bootstrapping). Must be
\code{<length(xx)}, so that a permutation is not composed of repetitions of only 1 element.}
}
\description{
Permute sample labels or indices (\code{xx}) into \code{nperm} matrix rows of length 
\code{length(xx)} for permutation testing. The rows are not duplicates of each other, 
and do not contain \code{xx}. Calculations are done with \pkg{arrangements}.
}
