% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph2kernel.R
\name{graph2kernel}
\alias{graph2kernel}
\title{Transform an \code{igraph} graph into a p-step random walk kernel}
\usage{
graph2kernel(gr, a = 2, p = 1)
}
\arguments{
\item{gr}{graph object of class \code{igraph}.}

\item{a}{parameter of p-step random walk kernel.}

\item{p}{parameter of p-step random walk kernel.}
}
\value{
Laplacian kernel matrix.
}
\description{
Transforms an igraph input graph into a p-step random walk kernel, using parameters \code{a} and \code{p}.
This is similar to \code{netClass::calc.diffusionKernelp}, except that function takes
slightly different input. The equation for the kernel is (a * I - L)^p.
}
