% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_ecdf.R
\name{p_ecdf}
\alias{p_ecdf}
\title{Estimate p-value from simulations}
\usage{
p_ecdf(eval.v, score.mat, alternative = c("two.sided", "less", "greater"))
}
\arguments{
\item{eval.v}{A vector of values where the cumulative distribution function should be evaluated.}

\item{score.mat}{A matrix of values from the distribution, where each column is a different simulation, 
and \code{rownames(score.mat)} corresponds to \code{names(eval.v)}.}

\item{alternative}{A character string specifying the alternative hypothesis.}
}
\value{
A matrix with two columns containing z-scores (larger is more significant) & p-values with 
\code{nrow = length(eval.v)}.
}
\description{
Estimate p-value by comparing a score to its simulations.
}
