% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_smooth_ecdf.R
\name{p_smooth_ecdf}
\alias{p_smooth_ecdf}
\title{Estimate p-value by smoothing density}
\usage{
p_smooth_ecdf(eval.point, scores, lower.tail = TRUE)
}
\arguments{
\item{eval.point}{A value where the cumulative distribution function should be evaluated}

\item{scores}{A vector of values from the distribution}

\item{lower.tail}{Logical; if TRUE (default), probabilities are P[X ≤ x] otherwise, P[X > x].}
}
\description{
Smooth a vector using a kernel density estimator then estimate a p-value from that density
}
