% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pants.R
\name{pants}
\alias{pants}
\title{Pathway analysis via network smoothing (Pants) testing group contrasts}
\usage{
pants(object, phenotype, contrast.v = NULL, mediation = FALSE,
  exposure = NULL, Gmat, covariates = NULL, ker = NULL,
  feat.tab = NULL, ntop = 25, score_fcn = abs, nperm = 10^4 - 1,
  ret.pwy.dfs = FALSE, ret.null.mats = FALSE, min.nfeats = 3,
  ncores = 1, name = NA, seed = 0)
}
\arguments{
\item{object}{Matrix-like data object containing log-ratios or log-expression values, with rows corresponding to 
features (e.g. genes) and columns to samples. Must have rownames that are non-duplicated and non-empty.}

\item{phenotype}{Vector of phenotypes of the samples. Should be same length as \code{ncol(object)}.}

\item{contrast.v}{Named vector of contrasts, passed to \code{\link[limma]{makeContrasts}}.}

\item{mediation}{Logical; should \code{\link[ezlimma]{hitman}} be applied?}

\item{exposure}{A numeric vector or matrix of exposures. Ignored if \code{hitman} is \code{FALSE}.}

\item{Gmat}{Binary feature (e.g. gene) by pathway inclusion matrix, indicating which features are in which pathways.}

\item{covariates}{Numeric vector with one element per sample or matrix-like object with rows corresponding 
to samples and columns to covariates to be adjusted for.}

\item{ker}{Laplacian kernel matrix representing the interaction network.}

\item{feat.tab}{Table of feature (e.g. gene) statistics and annotation that the Excel table can link to. If 
\code{NULL}, statistics from permutation testing are used.}

\item{ntop}{Number of top features that most impact a pathway to include.}

\item{score_fcn}{Function that transforms the t-statistics from the contrasts into a non-negative value. 
Its input must be a vector of same length as number of elements in \code{contrast.v} (usually one). 
Its output must be a non-negative scalar. Ignored if \code{hitman} is \code{TRUE}.}

\item{nperm}{Number of sample permutations to evaluate significance of pathways.}

\item{ret.pwy.dfs}{Logical; return list of data frames written out to CSVs?}

\item{ret.null.mats}{Logical; return matrices with null distributions for features and pathways?}

\item{min.nfeats}{Minimum number of features (e.g. genes) needed in a gene set for testing.}

\item{ncores}{Number of cores to use for parallel computing. You can detect how many are available for your system
using \code{\link[parallel]{detectCores}}.}

\item{name}{Name for the folder and Excel file that get written. Set to \code{NA} to avoid writing output.}

\item{seed}{Integer seed to set for reproducility.}
}
\value{
List of at least two data frames:
\describe{
   \item{\code{pwy.stats}}{A data frame with columns 
   \describe{
   \item{\code{nfeatures}}{number of features in the pathway.}
   \item{\code{score}}{only returned if \code{ret.null.mats} is \code{TRUE}; 
   pathway score (larger is more significant) to compare to \code{null.pwy.mat}}
   \item{\code{z}}{pathway permutation z-score (larger is more significant)}
   \item{\code{p}}{pathway permutation p-value}
   \item{\code{FDR}}{pathway FDR calculated from p-values with \code{p.adjust(p, method="BH")}}
   }}
   
   \item{\code{feature.stats}}{A data frame with columns
   \describe{
   \item{\code{score}}{without \code{mediation}, feature's score from applying \code{score_fcn} to moderated t-statistics;
   or with \code{mediation}, parametric z-score from \code{hitman}'s mediation p-value.}
   \item{\code{z}}{feature non-parametric z-score (larger is more significant) from comparing \code{score} vs. 
   this feature's scores in permutations (before smoothing)}
   \item{\code{p}}{feature's non-parametric permutation p-value}
   \item{\code{FDR}}{feature's non-parametric FDR from permutation \code{p}}
   }}
   
   if \code{ret.pwy.dfs} is \code{TRUE}:
   \item{\code{pwy.dfs}}{List of data frames written out to CSVs}
   
   And if \code{ret.null.mats} is \code{TRUE}:
   \item{\code{null.feature.mat}}{Matrix with features as rows and permutations as columns, where each element represents
   the score of that feature in that permutation}
   \item{\code{null.pwy.mat}}{Matrix with pathways as rows and permutations as columns, where each element represents
   the score of that pathway in that permutation}
   \item{\code{sample.perms}}{Matrix with samples as rows and permutations as columns, where each element represents
   the index of the sample simulated to represent the sample in the row in that permutation}
 }
}
\description{
Pants algorithm to test if group differences for features (i.e. analytes such as a gene, protein, or metabolite) 
in a pathway or those connected to the pathway in an interaction network are greater than randomized ones.
}
\details{
Without \code{mediation}, \code{phenotype}'s are permuted, since this properly permutes the \code{object} to 
\code{phenotype} mapping. \code{object} could be equivalently permuted. 
With \code{mediation},  because \code{object} is tested for its association to both \code{phenotype} 
and \code{exposure}, \code{colnames(object)} are permuted, which offers more available permutations.

Scores for features in the kernel but not in the data are assigned a z-score of zero by default for sparsity.
Scores for features and pathways are compared to null scores, which are generated by permuting the columns of 
\code{object} and rerunning the algorithm. These are the stats returned in \code{feature.stats}. However, \code{pants}
accepts \code{feat.tab} with statistics from \code{ezlimma} per-feature analysis for the Excel file to link to;
these may be more precise, since they don't rely on a limited number of permutations. If \code{feat.tab} is \code{NULL},
the Excel file links to tables derived from \code{feature.stats}.

For \code{\link[parallel]{makeCluster}}, the cluster \code{type} depends on the OS, which is tested in the body
of the function using \code{.Platform$OS.type}.

If \code{!is.na(name)}, an Excel file with "_pants.xlsx" appended to \code{name} gets written out with links to CSVs 
containing the statistics and annotation of features most affecting the pathway's score. The statistics and annotation
are from \code{feat.tab}, which is usually calculated with \pkg{ezlimma}. Additionally, the CSVs contain whether each 
feature is in the pathway, and an \code{impact} column describing the impact of each feature on the pathway's 
score. Since a pathway's score is calculated in \code{pants}, \code{impact} uses the feature statistics calculated 
in \code{pants} by comparing to permutation. The feature statistics from \pkg{ezlimma} and those from 
\code{pants} are nearly identical, though; the main difference is that \code{pants} feature significances are limited 
by the number of permutations, so they flatten near the extreme. The features with the largest magnitude impact score
are selected and can be visualized with \code{ezlimmaplot::plot_pwy}. These features may increase or decrease a 
pathway's score.
}
\examples{
# A workflow is described in the vignette; instructions to view the vignette are in the README.
}
