% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pants.R
\name{pants}
\alias{pants}
\title{Pathway Analysis via NeTwork Smoothing}
\usage{
pants(object, phenotypes.v, contrasts.v, ker, Gmat, score_fcn = abs,
  nperm = 10^4, smooth.pwy.null = NA, smooth.feat.null = NA,
  ret.null.mats = FALSE, verbose = FALSE)
}
\arguments{
\item{object}{A matrix-like data object containing log-ratios or log-expression values for a
series of samples, with rows corresponding to features and columns to samples.}

\item{phenotypes.v}{A vector of phenotypes of strings the same length as number of samples in \code{object}.
If the vector is named, the names must match the column names of \code{object}.}

\item{contrasts.v}{A named vector of constrasts. The constrasts must refer to the phenotypes
in \code{phenotypes.v}. Their order defines the order they are passed to \code{score_fcn}.}

\item{ker}{The Laplacian kernel matrix.}

\item{Gmat}{The feature by pathway inclusion matrix, indicating which features are in which pathways.}

\item{score_fcn}{A function that transforms the t-statistics from the contrasts.
Set to \code{identity} so it is the trivial identity function returning its argument.
Default is absolute value. Its input must be a vector of same length as number of elements in \code{contrasts.v}.
Its output must be a scalar (i.e. a vector of length one).}

\item{nperm}{Number of permutations to perform to evaluate significance of pathways.}

\item{smooth.pwy.null}{One of "kde" for kernel density estimation, "norm" for the normal distribution, or NA for 
no smoothing of null pathway statistics.}

\item{smooth.feat.null}{One of "kde" for kernel density estimation, "norm" for the normal distribution, or NA for 
no smoothing of null fature statistics.
#'@param ret.null.mats If TRUE, return matrices with null distributions for features and pathways.}

\item{verbose}{Logical indicating if the permutation number should be output every 100 permutations.}
}
\value{
List of length two, with elements \code{pwy.stats} and \code{feature.stats}, each containing a data frame
of statistics.
}
\description{
This is the main function of the \code{PANTS} package, which implements the PANTS algorithm
of Pathway Analysis (and optionally integration) via NeTwork Smoothing for group contrasts.
}
\details{
Scores for features in the kernel but not in the data are assigned a value of zero by default for sparsity.
Scores for features and pathways are compared to null scores, which are generated by permuting the phenotypes and 
rerunning the algorithm. These null scores can be smoothed using \code{\link{density}} or fit to a known distribution,
such as the normal or exponential distribution.
}
