% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pants.R
\name{pants}
\alias{pants}
\title{Pathway analysis via network smoothing}
\usage{
pants(object, phenotype, contrast.v, Gmat, ker = NULL, annot = NULL,
  score_fcn = identity, nperm = 10^4 - 1, ret.null.mats = FALSE,
  alternative = c("two.sided", "less", "greater"), min.nfeats = 0,
  ncores = 1, name = NA, n.toptabs = Inf, seed = 0)
}
\arguments{
\item{object}{Matrix-like data object containing log-ratios or log-expression values, with rows corresponding to 
features (e.g. genes) and columns to samples. Must have rownames that are non-duplicated and non-empty.}

\item{phenotype}{Vector of phenotypes of the samples. Should be same length as \code{ncol(object)}.}

\item{contrast.v}{Named vector of contrasts, passed to \code{\link[limma]{makeContrasts}}.}

\item{Gmat}{Binary feature (e.g. gene) by pathway inclusion matrix, indicating which features are in which pathways.}

\item{ker}{Laplacian kernel matrix.}

\item{annot}{Data frame of feature annotation.}

\item{score_fcn}{A function that transforms the t-statistics from the contrasts. \code{identity} is 
the trivial identity function returning its argument. Its input must be a vector of same 
length as number of elements in \code{contrast.v}. Its output must be a scalar.}

\item{nperm}{Number of permutations to perform to evaluate significance of pathways.}

\item{ret.null.mats}{If TRUE, return matrices with null distributions for features and pathwaysle}

\item{alternative}{Alternative hypothesis; must be one of \code{"two.sided"}, \code{"greater"} or \code{"less"}. 
\code{"greater"} corresponds to positive association, \code{"less"} to negative association.}

\item{min.nfeats}{Minimum number of features (e.g. genes) needed in a gene set for testing.}

\item{ncores}{Number of cores to use for parallel computing. You can detect how many are available for your system
using \code{\link[parallel]{detectCores}}.}

\item{name}{Name for the folder and Excel file that get written. Set to \code{NA} to avoid writing output.}

\item{n.toptabs}{Number of gene set toptables to write to CSV and link to from Excel.}

\item{seed}{Integer seed to set for reproducility if \code{fun="mroast"}, since \code{mroast} uses random 
simulations. Ignored if \code{fun="fry"}.}
}
\value{
List of at least two dataframes:
\describe{
   \item{pwy.stats}{A dataframe with columns 
   \describe{
   \item{nfeatures}{number of features in the pathway} 
   \item{feat.score.avg}{sum of smoothed scores of the pathway's features / \code{nfeatures}. This score is compared
   to scores in permutations. Only included if \code{ret.null.mats==TRUE}.}
   \item{z}{pathway permutation z-score (larger is more significant)}
   \item{p}{pathway permutation p-value} 
   \item{FDR}{pathway FDR calculated from p-values with \code{p.adjust(p, method='BH')}}
   }}
   \item{feature.stats}{A dataframe with columns
   \describe{
   \item{score}{feature's score from applying \code{score_fcn} in \code{\link{score_features}}}
   \item{z}{feature z-score (larger is more significant) relative to this feature's scores in permutation 
   (without smoothing)} 
   \item{p}{feature's permutation p-value} 
   \item{FDR}{feature's FDR from permutation \code{p}}
   }}
   And if \code{ret.null.mats} is TRUE:
   \item{null.feature.mat}{Matrix with features as rows and permutations as columns, where each element represents
   the score of that feature in that permutation}
   \item{null.pwy.mat}{Matrix with pathways as rows and permutations as columns, where each element represents
   the score of that pathway in that permutation}
 }
}
\description{
Implements the Pants algorithm for pathway analysis via network smoothing with group contrasts.
}
\details{
Scores for features in the kernel but not in the data are assigned a z-score of zero by default for sparsity.
Scores for features and pathways are compared to null scores, which are generated by permuting the columns of 
\code{object} and rerunning the algorithm.

For \code{\link[parallel]{makeCluster}}, the cluster \code{type} depends on the OS, which is tested in the body
of the function using \code{.Platform$OS.type}.

If \code{!is.na(name)}, an Excel file gets written out with links to toptable CSVs, containing the statistics of features in
the pathway. If \code{ker} is used, pathways may be affected via smoothing by features outside the pathway. These can be seen
with \code{\link[PANTS]{plot_pwy}}.
}
