% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pants.R
\name{pants}
\alias{pants}
\title{Pathway analysis via network smoothing (Pants) testing group contrasts}
\usage{
pants(object, phenotype, contrast.v, Gmat, ker = NULL, feat.tab = NULL,
  ntop = 25, score_fcn = identity, nperm = 10^4 - 1,
  ret.null.mats = FALSE, alternative = c("two.sided", "less",
  "greater"), min.nfeats = 3, ncores = 1, name = NA, seed = 0)
}
\arguments{
\item{object}{Matrix-like data object containing log-ratios or log-expression values, with rows corresponding to 
features (e.g. genes) and columns to samples. Must have rownames that are non-duplicated and non-empty.}

\item{phenotype}{Vector of phenotypes of the samples. Should be same length as \code{ncol(object)}.}

\item{contrast.v}{Named vector of contrasts, passed to \code{\link[limma]{makeContrasts}}.}

\item{Gmat}{Binary feature (e.g. gene) by pathway inclusion matrix, indicating which features are in which pathways.}

\item{ker}{Laplacian kernel matrix representing the interaction network.}

\item{feat.tab}{Table of feature (e.g. gene) statistics and annotation that the Excel table can link to. If 
\code{NULL}, statistics from permutation testing are used.}

\item{ntop}{Number of top features that most impact a pathway to include.}

\item{score_fcn}{A function that transforms the t-statistics from the contrasts. \code{identity} is 
the trivial identity function returning its argument. Its input must be a vector of same 
length as number of elements in \code{contrast.v}. Its output must be a scalar.}

\item{nperm}{Number of sample permutations of phenotype to evaluate significance of pathways.}

\item{ret.null.mats}{If TRUE, return matrices with null distributions for features and pathwaysle}

\item{alternative}{Alternative hypothesis; must be one of \code{"two.sided"}, \code{"greater"} or \code{"less"}. 
\code{"greater"} corresponds to positive association, \code{"less"} to negative association.}

\item{min.nfeats}{Minimum number of features (e.g. genes) needed in a gene set for testing.}

\item{ncores}{Number of cores to use for parallel computing. You can detect how many are available for your system
using \code{\link[parallel]{detectCores}}.}

\item{name}{Name for the folder and Excel file that get written. Set to \code{NA} to avoid writing output.}

\item{seed}{Integer seed to set for reproducility if \code{fun="mroast"}, since \code{mroast} uses random 
simulations. Ignored if \code{fun="fry"}.}
}
\value{
List of at least two data frames:
\describe{
   \item{\code{pwy.stats}}{A data frame with columns 
   \describe{
   \item{\code{nfeatures}}{number of features in the pathway.}
   \item{\code{z}}{pathway permutation z-score (larger is more significant)}
   \item{\code{p}}{pathway permutation p-value}
   \item{\code{FDR}}{pathway FDR calculated from p-values with \code{p.adjust(p, method="BH")}}
   }}
   \item{\code{feature.stats}}{A data frame with columns
   \describe{
   \item{\code{score}}{feature's score from applying \code{score_fcn} in \code{\link{score_features}}}
   \item{\code{z}}{feature z-score (larger is more significant) relative to this feature's scores in permutation 
   (without smoothing)} 
   \item{\code{p}}{feature's permutation p-value}
   \item{\code{FDR}}{feature's FDR from permutation \code{p}}
   }}
   And if \code{ret.null.mats} is \code{TRUE}:
   \item{\code{null.feature.mat}}{Matrix with features as rows and permutations as columns, where each element represents
   the score of that feature in that permutation}
   \item{\code{null.pwy.mat}}{Matrix with pathways as rows and permutations as columns, where each element represents
   the score of that pathway in that permutation}
   \item{\code{sample.perms}}{Matrix with samples as rows and permutations as columns, where each element represents
   the index of the sample simulated to represent the sample in the row in that permutation}
 }
}
\description{
Pants algorithm to test group differences for features (i.e. analytes such as a gene, protein, or metabolite) in a 
pathway or those connected to the pathway in an interaction network.
}
\details{
Scores for features in the kernel but not in the data are assigned a z-score of zero by default for sparsity.
Scores for features and pathways are compared to null scores, which are generated by permuting the columns of 
\code{object} and rerunning the algorithm. These are the stats returned in \code{feature.stats}. However, \code{pants}
accepts \code{feat.tab} with statistics from \code{ezlimma} per-feature analysis for the Excel file to link to;
these may be more precise, since they don't rely on a limited number of permutations. If \code{feat.tab} is \code{NULL},
the Excel file links to tables derived from \code{feature.stats}.

For \code{\link[parallel]{makeCluster}}, the cluster \code{type} depends on the OS, which is tested in the body
of the function using \code{.Platform$OS.type}.

If \code{!is.na(name)}, an Excel file with "_pants.xlsx" appended to \code{name} gets written out with links to CSVs 
containing the statistics and annotation of features most affecting the pathway's score. The statistics and annotation
are from \code{feat.tab}, which is usually calculated with \pkg{ezlimma}. Additionally, the CSVs contain whether each 
feature is in the pathway, and an \code{impact} column describing the impact of each feature on the pathway's 
score. Since a pathway's score is calculated in \code{pants}, \code{impact} uses the feature statistics calculated 
in \code{pants} by comparing to permutation. The feature statistics from \pkg{ezlimma} and those from 
\code{pants} are nearly identical, though; the main difference is that \code{pants} feature significances are limited 
by the number of permutations, so they flatten near the extreme. The features with the largest magnitude impact score
are selected and can be visualized with \code{\link[PANTS]{plot_pwy}}. Whether these features increase or decrease a 
pathway's score depends on the \code{alternative}.
}
\examples{
# A workflow is described in the vignette; instructions to view the vignette are in the README.
}
