% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pants_hitman.R
\name{pants_hitman}
\alias{pants_hitman}
\title{Pathway Analysis via network smoothing using Hitman}
\usage{
pants_hitman(object, exposure, phenotype.v, Gmat, covariates = NULL,
  ker = NULL, annot = NULL, nperm = 10^4 - 1,
  ret.null.mats = FALSE, min.size = 0, ncores = 1, name = NA,
  n.toptabs = Inf, seed = 0)
}
\arguments{
\item{object}{A matrix-like data object containing log-ratios or log-expression values for a
series of samples, with rows corresponding to features and columns to samples.}

\item{exposure}{A numeric vector or matrix of exposures.}

\item{phenotype.v}{A vector of numeric phenotypes the same length as number of samples in \code{object}. If the 
vector is named, the names must match the column names of \code{object}.}

\item{Gmat}{Binary feature by pathway inclusion matrix, indicating which features are in which pathways.}

\item{covariates}{Numeric vector or matrix of covariates.}

\item{ker}{Laplacian kernel matrix.}

\item{annot}{A data frame of feature annotation.}

\item{nperm}{Number of permutations to perform to evaluate significance of pathways.}

\item{ret.null.mats}{If TRUE, return matrices with null distributions for features and pathways.}

\item{min.size}{Pathways with fewer than \code{min.size} measured features in \code{object} are filtered out.}

\item{ncores}{Number of cores to use for parallel computing. You can detect how many are available for your system
using \code{\link[parallel]{detectCores}}.}

\item{name}{Name for the folder and Excel file that get written. Set to \code{NA} to avoid writing output.}

\item{n.toptabs}{number of pathway toptables to write to CSV and link to from Excel.}

\item{seed}{Seed to set using \code{set.seed} for reproducibility.}
}
\value{
List of at least two dataframes:
\describe{
   \item{pwy.stats}{A dataframe with columns 
   \describe{
   \item{nfeatures}{number of features in the pathway} 
   \item{feat.score.avg}{sum of smoothed scores of the pathway's features / \code{nfeatures}. This score is compared
   to scores in permutations. Only included if \code{ret.null.mats==TRUE}.}
   \item{z}{pathway permutation z-score (larger is more significant)}
   \item{p}{pathway permutation p-value} 
   \item{FDR}{pathway FDR calculated from p-values with \code{p.adjust(p, method='BH')}}
   }}
   \item{feature.stats}{A dataframe with columns
   \describe{
   \item{score}{feature's score, which is p-value from \code{\link[ezlimma]{hitman}} transformed into z-score 
   (larger is more significant)}
   \item{z}{feature z-score (larger is more significant) relative to this feature's scores in permutation 
   without smoothing. This should be similar to \code{score} if the permutation null distribution matches the 
   theoretical null distribution.} 
   \item{p}{feature's permutation p-value} 
   \item{FDR}{feature's FDR from permutation \code{p}}
   }}
   And if \code{ret.null.mats} is TRUE:
   \item{null.feature.mat}{Matrix with features as rows and permutations as columns, where each element represents
   the score of that feature in that permutation}
   \item{null.pwy.mat}{Matrix with pathways as rows and permutations as columns, where each element represents
   the score of that pathway in that permutation}
 }
}
\description{
Pants algorithm for pathway mediation analysis with Hitman and network smoothing.
}
\details{
Scores for features in the kernel but not in the data are assigned a z-score of zero by default for sparsity.
Scores for features and pathways are compared to null scores, which are generated by permuting the columns of 
\code{object} and rerunning the algorithm.

For \code{\link[parallel]{makeCluster}}, the cluster \code{type} depends on the OS, which is tested in the body
of the function using \code{.Platform$OS.type}.

If \code{!is.na(name)}, an Excel file gets written out with links to toptable CSVs, containing the statistics of features in
the pathway. If \code{ker} is used, pathways may be affected via smoothing by features outside the pathway. These can be seen
with \code{\link[PANTS]{plot_pwy}}.
}
