% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pants_hitman.R
\name{pants_hitman}
\alias{pants_hitman}
\title{Pathway Analysis via Network Smoothing using Hitman}
\usage{
pants_hitman(object, exposure, phenotype.v, Gmat, covariates = NULL,
  ker = NULL, nperm = 10^4, ret.null.mats = FALSE, verbose = TRUE,
  min.size = 0, seed = 0)
}
\arguments{
\item{object}{A matrix-like data object containing log-ratios or log-expression values for a
series of samples, with rows corresponding to features and columns to samples.}

\item{exposure}{A numeric vector or matrix of exposures.}

\item{phenotype.v}{A vector of numeric phenotypes the same length as number of samples in \code{object}. If the 
vector is named, the names must match the column names of \code{object}.}

\item{Gmat}{Binary feature by pathway inclusion matrix, indicating which features are in which pathways.}

\item{covariates}{Numeric vector or matrix of covariates.}

\item{ker}{Laplacian kernel matrix.}

\item{nperm}{Number of permutations to perform to evaluate significance of pathways.}

\item{ret.null.mats}{If TRUE, return matrices with null distributions for features and pathways.}

\item{verbose}{Logical indicating if the permutation number should be output every 500 permutations.}

\item{min.size}{Pathways with fewer than \code{min.size} measured features in \code{object} are filtered out.}

\item{seed}{Seed to set using \code{set.seed} for reproducibility.}
}
\value{
List of at least two dataframes:
\describe{
   \item{pwy.stats}{A dataframe with columns 
   \describe{
   \item{nfeatures}{number of features in the pathway} 
   \item{feat.score.avg}{sum of smoothed scores of the pathway's features / \code{nfeatures}. This score is compared
   to scores in permutations.}
   \item{z}{pathway permutation z-score (larger is more significant)}
   \item{p}{pathway permutation p-value} 
   \item{FDR}{pathway FDR calculated from p-values with \code{p.adjust(p, method='BH')}}
   }}
   \item{feature.stats}{A dataframe with columns
   \describe{
   \item{score}{feature's score, which is p-value from \code{\link[ezlimma]{hitman}} transformed into z-score 
   (larger is more significant)}
   \item{z}{feature z-score (larger is more significant) relative to this feature's scores in permutation 
   without smoothing. This should be similar to \code{score} if the permutation null distribution matches the 
   theoretical null distribution.} 
   \item{p}{feature's permutation p-value} 
   \item{FDR}{feature's FDR from permutation \code{p}}
   }}
   And if \code{ret.null.mats} is TRUE:
   \item{null.feature.mat}{Matrix with features as rows and permutations as columns, where each element represents
   the score of that feature in that permutation}
   \item{null.pwy.mat}{Matrix with pathways as rows and permutations as columns, where each element represents
   the score of that pathway in that permutation}
 }
}
\description{
Pants algorithm for pathway mediation analysis with Hitman and network smoothing.
}
\details{
Scores for features in the kernel but not in the data are assigned a z-score of zero by default for sparsity.
Scores for features and pathways are compared to null scores, which are generated by permuting the columns of 
\code{object} and rerunning the algorithm.
}
