% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pants_mediation.R
\name{pants_mediation}
\alias{pants_mediation}
\title{Pathway Analysis via Network Smoothing using mediation}
\usage{
pants_mediation(object, exposure.v, phenotype.v, ker, Gmat, nperm = 10^4,
  ret.null.mats = FALSE, verbose = TRUE)
}
\arguments{
\item{object}{A matrix-like data object containing log-ratios or log-expression values for a
series of samples, with rows corresponding to features and columns to samples.}

\item{exposure.v}{A numeric or character vector of exposures the same length as number of samples in \code{object}.
If the vector is named, the names must match the column names of \code{object}.}

\item{phenotype.v}{A vector of numeric phenotypes the same length as number of samples in \code{object}. If the 
vector is named, the names must match the column names of \code{object}.}

\item{ker}{The Laplacian kernel matrix.}

\item{Gmat}{The feature by pathway inclusion matrix, indicating which features are in which pathways.}

\item{nperm}{Number of permutations to perform to evaluate significance of pathways.}

\item{ret.null.mats}{If TRUE, return matrices with null distributions for features and pathways.}

\item{verbose}{Logical indicating if the permutation number should be output every 500 permutations.}
}
\value{
List of at least two dataframes:
\describe{
   \item{pwy.stats}{A dataframe with columns 
   \describe{
   \item{nfeatures}{number of features in the pathway} 
   \item{feat.score.avg}{sum of smoothed scores of the pathway's features / \code{nfeatures}}
   \item{z}{pathway permutation z-score (larger is more significant)}
   \item{p}{pathway permutation p-value} 
   \item{FDR}{pathway FDR calculated from p-values with \code{p.adjust(p, method='BH')}}
   }}
   \item{feature.stats}{A dataframe with columns
   \describe{
   \item{score}{feature's score from \code{\link{score_features}}}
   \item{z}{feature z-score (larger is more significant) relative to this feature's scores in permutation 
   (without smoothing)} 
   \item{p}{feature's permutation p-value} 
   \code{FDR}{feature's FDR from permutation \code{p}}
   }}
   And if \code{ret.null.mats} is TRUE:
   \item{null.feature.mat}{Matrix with features as rows and permutations as columns, where each element represents
   the score of that feature in that permutation}
   \item{null.pwy.mat}{Matrix with pathways as rows and permutations as columns, where each element represents
   the score of that pathway in that permutation}
 }
}
\description{
PANTS algorithm for mediation pathway analysis (and optionally integration) via network smoothing
}
\details{
Scores for features in the kernel but not in the data are assigned a z-score of zero by default for sparsity.
Scores for features and pathways are compared to null scores, which are generated by permuting the columns of 
\code{object} and rerunning the algorithm.
}
