% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pwy.R
\name{plot_pwy}
\alias{plot_pwy}
\title{Plot network diagram for a pathway}
\usage{
plot_pwy(gr, score.v, ntop = 7, Gmat, pwy, ker = NULL, annot.v = NA,
  alternative = c("two.sided", "less", "greater"), name = NULL,
  plot = TRUE, signif.dig = 2, seed = 0)
}
\arguments{
\item{gr}{graph object of class \code{igraph}.}

\item{score.v}{Named vector of scores of features to select high impact nodes.}

\item{ntop}{Number of top features that most impact a pathway to include.}

\item{Gmat}{Binary feature (e.g. gene) by pathway inclusion matrix, indicating which features are in which pathways.}

\item{pwy}{Pathway, must be a column name of \code{Gmat}.}

\item{ker}{Laplacian kernel matrix representing the interaction network.}

\item{annot.v}{Named vector of annotations for nodes. If \code{annot.v} is given, \code{names(annot.v)} should 
have some overlap with \code{rownames(Gmat)}.}

\item{alternative}{Alternative hypothesis; must be one of \code{"two.sided"}, \code{"greater"} or \code{"less"}. 
\code{"greater"} corresponds to positive association, \code{"less"} to negative association.}

\item{name}{Name of file to plot to. If \code{NULL}, creates a filename from \code{pwy} and \code{ntop}. Set to 
\code{NA} to plot to screen instead of to file.}

\item{plot}{Logical; should plot be generated?}

\item{signif.dig}{Number of significant digits to use for colorbar axis labels.}

\item{seed}{Integer seed to set for reproducility if \code{fun="mroast"}, since \code{mroast} uses random 
simulations. Ignored if \code{fun="fry"}.}
}
\value{
Invisibly, a list with components: 
 \describe{
   \item{\code{gr}}{the graph that gets plotted}
   \item{\code{vertex.color}}{the vertex colors}
   \item{\code{vertex.shape}}{the vertex shapes}
   \item{\code{score}}{scores of the vertices of the plotted graph}
   \item{\code{top.nodes}}{ntop top nodes driving pathway}
   \item{\code{impact}}{impact score on pathways}
}
}
\description{
Plot nodes most impacting a pathway as a network diagram with node color corresponding to significance and 
shape to pathway membership. The most impactful nodes are inferred by assuming the input here is the same as
was used to calculate pathway significance in \code{\link{pants}} or \code{\link{pants_hitman}}.
}
\details{
If some \code{ntop} nodes are outside \code{pwy}, then their neighbor nodes inside \code{pwy} are also 
plotted. These nodes are then connected based on the interaction network. Nodes inside the network are drawn as
circles, whereas those outside the network are drawn as squares, which is indicated by a legend when some nodes
are outside the pathway.

Unmeasued nodes have \code{is.na(score.v)} and are drawn as white, which is indicated in the colorbar.

The color palette is from \code{\link[RColorBrewer]{brewer.pal}}, and depends on the \code{alternative}.
}
