% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pwy.R
\name{plot_pwy}
\alias{plot_pwy}
\title{Plot network diagram for a pathway}
\usage{
plot_pwy(gr, score.v, ntop = 7, Gmat, pwy, ker = NULL, annot.v = NA,
  alternative = c("two.sided", "less", "greater"), name = NULL,
  color.pal = NULL, plot = TRUE, seed = 0)
}
\arguments{
\item{gr}{graph object of class \code{igraph}.}

\item{score.v}{Named vector of scores of features to select driver nodes.}

\item{ntop}{Number of top features that most impact a pathway to include.}

\item{Gmat}{Binary feature (e.g. gene) by pathway inclusion matrix, indicating which features are in which pathways.}

\item{pwy}{Pathway, must be a column name of \code{Gmat}.}

\item{ker}{Laplacian kernel matrix representing the interaction network.}

\item{annot.v}{Named vector of annotations for nodes. If \code{annot.v} is given, \code{names(annot.v)} should 
have some overlap with \code{rownames(Gmat)}.}

\item{alternative}{Alternative hypothesis; must be one of \code{"two.sided"}, \code{"greater"} or \code{"less"}. 
\code{"greater"} corresponds to positive association, \code{"less"} to negative association.}

\item{name}{Name of file to create. Set to \code{NA} to plot to screen instead of to file.}

\item{color.pal}{A color palette, as a vector. Must be accepted by \code{\link[igraph]{plot.igraph}}. If \code{NULL},
a palette from \code{\link[RColorBrewer]{brewer.pal}} appropriate to \code{alternative} is chosen.}

\item{plot}{Logical, should plot be generated?}

\item{seed}{Integer seed to set for reproducility if \code{fun="mroast"}, since \code{mroast} uses random 
simulations. Ignored if \code{fun="fry"}.}
}
\value{
Invisibly, a list with components: 
 \describe{
   \item{\code{gr}}{the graph that gets plotted}
   \item{\code{vertex.color}}{the vertex colors}
   \item{\code{vertex.shape}}{the vertex shapes}
   \item{\code{score}}{scores of the vertices of the plotted graph}
   \item{\code{top.nodes}}{ntop top nodes driving pathway}
   \item{\code{impact}}{impact score on pathways}
}
}
\description{
Plot driver nodes for pathway \code{pwy} as a network diagram with node color corresponding to significance and 
shape to pathway membership. The driver nodes are inferred by assuming the input here is the same as was used 
to calculate pathway significance in \code{\link{pants}} or \code{\link{pants_hitman}}.
}
\details{
It's checked that \code{names(score.v) == rownames(gr)}. If some \code{ntop} nodes are external, then their 
internal neighbor nodes are also included. These nodes are then connected based on the interaction network.
}
