% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_features.R
\name{score_features}
\alias{score_features}
\title{Score features using ezlimma and score_fcn}
\usage{
score_features(object, phenotypes.v, contrasts.v, ker, Gmat, score_fcn)
}
\arguments{
\item{object}{A matrix-like data object containing log-ratios or log-expression values for a
series of samples, with rows corresponding to features and columns to samples.}

\item{phenotypes.v}{A vector of phenotypes of strings the same length as number of samples in \code{object}.
If the vector is named, the names must match the column names of \code{object}.}

\item{contrasts.v}{A named vector of constrasts. The constrasts must refer to the phenotypes
in \code{phenotypes.v}. Their order defines the order they are passed to \code{score_fcn}.}

\item{score_fcn}{A function that transforms the t-statistics from the contrasts.
Set to \code{identity} so it is the trivial identity function returning its argument.
Default is absolute value. Its input must be a vector of same length as number of elements in \code{contrasts.v}.
Its output must be a scalar (i.e. a vector of length one).}
}
\value{
Vector of feature scores.
}
\description{
Score features using ezlimma::limma_contrasts and score_fcn
}
