% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_features.R
\name{score_features}
\alias{score_features}
\title{Score features using ezlimma and score_fcn}
\usage{
score_features(object, phenotype, contrast.v, score_fcn = identity)
}
\arguments{
\item{object}{Matrix-like data object containing log-ratios or log-expression values, with rows corresponding to 
features (e.g. genes) and columns to samples. Must have rownames that are non-duplicated and non-empty.}

\item{phenotype}{Vector of phenotypes of the samples. Should be same length as \code{ncol(object)}.}

\item{contrast.v}{Named vector of contrasts, passed to \code{\link[limma]{makeContrasts}}.}

\item{score_fcn}{Function that transforms the t-statistics from the contrasts into a non-negative value. 
Its input must be a vector of same length as number of elements in \code{contrast.v} (usually one). 
Its output must be a non-negative scalar.}
}
\value{
Vector of feature scores.
}
\description{
Score features using ezlimma::limma_contrasts and score_fcn
}
