% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMPDB2Gmat.R
\name{SMPDB2Gmat}
\alias{SMPDB2Gmat}
\title{Transform SMPDB protein and/or metabolite pathway data frames into a pathway matrix}
\usage{
SMPDB2Gmat(smpdb.prot, smpdb.met, exclude.pwy.subj = NA)
}
\arguments{
\item{smpdb.prot}{Data frame of proteins per pathway from SMPDB.}

\item{smpdb.met}{Data frame of metabolites per pathway from SMPDB.}

\item{exclude.pwy.subj}{Vector of subjects (or types) of pathways to exclude}
}
\value{
A numeric Matrix of features (proteins and/or metabolites) as rows and pathways as columns, indicating 
pathway membership.
}
\description{
Transform SMPDB protein and/or metabolite pathway data frames, read in as CSV files from http://smpdb.ca/downloads,
into a pathway matrix that matches Pathway Commons, using Gene.Name to identify proteins and ChEBI IDs for metabolites.
Format of SMPDB downloads has changed, so now it has numerous semi-redundant fatty acid pathways, so this will need to
be modified if want to stay up to date with SMPDB.
}
