% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_smpdb_gmat.R
\docType{data}
\name{smpdb_gmat}
\alias{smpdb_gmat}
\title{Pathway inclusion matrix (Gmat) constructed from SMPDB.}
\format{A Matrix of features (proteins/metabolites) by Pathway names. Zero entries indicate the feature is not in the pathway,
whereas non-zero entries indicates otherwise.}
\usage{
smpdb_gmat
}
\description{
To construct this, I read in 'smpdb_metabolites.csv' and 'smpdb_proteins.csv' from the Small Molecule Pathway
Database (SMPDB), then SMPDB2Gmat(smpdb.prot=smpdb.prot, smpdb.met=smpdb.met, exclude.pwy.subj='Drug Metabolism').
}
\keyword{datasets}
