% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_gmat.R
\name{subset_gmat}
\alias{subset_gmat}
\title{Subset pathway inclusion matrix based on size}
\usage{
subset_gmat(object, Gmat, min.size)
}
\arguments{
\item{object}{A matrix-like data object containing log-ratios or log-expression values for a
series of samples, with rows corresponding to features and columns to samples.}

\item{Gmat}{Binary feature by pathway inclusion matrix, indicating which features are in which pathways.}

\item{min.size}{Pathways with fewer than \code{min.size} measured features in \code{object} are filtered out.}
}
\value{
List of two: \code{Gmat} and vector \code{nfeats.per.pwy} with number of features per pathway.
}
\description{
Subset pathway inclusion matrix based on number of measured features in pathway.
}
