% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_pants_xl.R
\name{write_pants_xl}
\alias{write_pants_xl}
\title{Write Excel XLSX file with links to CSVs}
\usage{
write_pants_xl(zscores, pwy.tab, feat.tab, Gmat, ker, name, ntop = 5)
}
\arguments{
\item{zscores}{Matrix with rownames as feature names, and 1st column with z-scores to select high impact nodes. 
Non-\code{NA} values should be finite.}

\item{pwy.tab}{A table of pathway results.}

\item{feat.tab}{Table of feature (e.g. gene) statistics and annotation.}

\item{Gmat}{Binary feature (e.g. gene) by pathway inclusion matrix, indicating which features are in which pathways.}

\item{ker}{Laplacian kernel matrix representing the interaction network.}

\item{name}{Name for the folder and Excel file that get written. Set to \code{NA} to avoid writing output.}

\item{ntop}{Number of top features that most impact a pathway to include.}
}
\value{
Invisibly, a list with two components:
\describe{
   \item{\code{pwys.xl}}{Data frame representing written Excel file}
   \item{\code{pwy.csvs}}{List of data frames representing each written pathway CSV file}
 }
}
\description{
Create directory; write CSV files with stats from \code{feat.tab} for all features in each pathway, and the top
features that impact each pathway (even if they're outside the pathway); and write an Excel XLSX file using 
\code{writexl} that links to the CSVs.
}
