% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_pants_xl.R
\name{write_pants_xl}
\alias{write_pants_xl}
\title{Write Excel XLSX file with links to CSVs}
\usage{
write_pants_xl(zscore.v, pwy.tab, feat.tab, Gmat, ker, name,
  alternative = c("two.sided", "less", "greater"), ntop = 5)
}
\arguments{
\item{zscore.v}{Named vector of feature z-scores to select high impact nodes. These should be finite unless they 
are \code{NA}.}

\item{pwy.tab}{A table of pathway results.}

\item{feat.tab}{Table of feature (e.g. gene) statistics and annotation that the Excel table can link to. If 
\code{NULL}, statistics from permutation testing are used.}

\item{Gmat}{Binary feature (e.g. gene) by pathway inclusion matrix, indicating which features are in which pathways.}

\item{ker}{Laplacian kernel matrix representing the interaction network.}

\item{name}{Name for the folder and Excel file that get written. Set to \code{NA} to avoid writing output.}

\item{alternative}{Alternative hypothesis; must be one of \code{"two.sided"}, \code{"greater"} or \code{"less"}. 
\code{"greater"} corresponds to positive association, \code{"less"} to negative association.}

\item{ntop}{Number of top features that most impact a pathway to include.}
}
\description{
Create directory; write CSV files with stats from \code{feat.tab} for all features in each pathway, and the top
features that impact each pathway (even if they're outside the pathway); and write an Excel XLSX file using 
\code{writexl} that links to the CSVs.
}
